+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/singularityhub/singularity-hpc/tree/main) (0.0.x)
+ - support for installing to symlink tree (0.0.49)
+  - also including cleanup of symlink tree on uninstall
+  - ability to set custom config variable on the fly with -c
  - Properly cleanup empty module directories, and asking to remove a container that doesn't exist now logs a _warning_ (0.0.48)
  - wrapper script generation permissions error (0.0.47)
  - fixing but with stream command repeating output (0.0.46)